+++ .gitignore
@@ -55,10 +55,14 @@ typings/
 .yarn-integrity
 
 # dotenv environment variables file
-.env 
+.env
 
 # next.js build output
 .next
 
 .vscode/*
+
+# cypress testing ignoring
+cypress/
+cypress.env.json
