+++ cypress.json
@@ -1,0 +1,3 @@
+{
+  "integrationFolder": "test/browser"
+}
