+++ test/browser/index.js
@@ -1,0 +1,5 @@
+import Anilist from '../../index';
+import addSuite from '../anilist.spec';
+
+const qaToken = Cypress.env('QA_TOKEN');
+addSuite(Anilist, qaToken);
